Added another Section
Used same photos but parsed out two of birds and renamed them.  Remaining "anim" jpgs renamed to have sequential numbering.  In TableViewController, added a new variable titled birdItems, added a new for-loop in the viewDidLoad method for the newly named image files and added a new append statement after the for-loop change