Descriptive cell title
Created descriptive named arrays to populate, then,  the DataItems file which then gives a name to match the image revealed.  Statements were added to the viewDidLoad override method.  All three start with a "let array name" to set the names.  To accommodate these statements, many of the files had to be renamed to remove any filename that had a numbering technique starting with a zero.